Added a DbWebHookStore abstraction which parameterizes DbContext and the registration classes currently used by SqlWebHookStore, This allows SqlWebHookStore as well as other DbContext and registration implementations to use DbWebHookStore. This can for example be used to enable support for Oracle and other databases supported by Entity Framework.
