+++ src/Microsoft.AspNet.WebHooks.Custom.SqlStorage/Properties/SqlStorageResources.resx
@@ -124,7 +124,7 @@
     <value>Concurrency failure during '{0}' operation: '{1}'.</value>
   </data>
   <data name="SqlStore_NoConnectionString" xml:space="preserve">
-    <value>Please provide an Azure Table Storage connection string with name '{0}' in the configuration string section of the 'Web.Config' file.</value>
+    <value>Please provide a SQL connection string with name '{0}' in the configuration string section of the 'Web.Config' file.</value>
   </data>
   <data name="SqlStore_OperationFailed" xml:space="preserve">
     <value>General error during '{0}' operation: '{1}'.</value>
